Some reorganizing and chngaes as per ticket HTDC-23442 and HTDC-23443
