+++ kmipcli/Makefile
