+++ kmipcli/cmd/KMIPCLIutils.go
