+++ kmipcli/cmd/disableKEK.go
