+++ kmipcli/cmd/getHSMInfo.go
