+++ kmipcli/cmd/getKmipObject.go
