+++ kmipcli/cmd/getKmipObjectCount.go
