+++ kmipcli/cmd/getPlatformInfo.go
