+++ kmipcli/cmd/listKmipObjects.go
