+++ kmipcli/cmd/locateRootKey.go
