+++ kmipcli/cmd/updateKmipObject.go
